+++ src/Core/Maths/Math.h
@@ -94,7 +94,7 @@ namespace CasaEngine
 	 */
 	inline float Sigmoid(float input, float response)
 	{
-		return 1.0f / (1.0f + static_cast<float>(exp(-input / response)));
+		return 1.0f / (1.0f + exp(-input / response));
 	}
 
 	//rounds a double up or down depending on its value
